feat: 레디스 설정 값 및 유저 기록값 타입 통일

+++ dailyquest_batch/src/main/kotlin/dailyquest/user/dto/UserPerfectDayCount.kt
@@ -5,7 +5,7 @@ import java.io.Serializable
 
 class UserPerfectDayCount(
     val userId: Long,
-    val perfectDayCount: Int
+    val perfectDayCount: Long
 ) : Serializable {
     companion object {
         @JvmStatic
